ai-service/requirements.txt
fastapi
uvicorn
python-multipart
easyocr
torch
transformers
PyMuPDF  # for standard PDF to image conversion (import fitz)
Pillow
pydantic
